+++ lib/shenandoah/locator.rb
@@ -5,13 +5,15 @@ module Shenandoah
     attr_accessor :main_path, :spec_path, :tmp_path
 
     def initialize(options = {})
-      @main_path = options[:main_path] || "lib"
-      @spec_path = options[:spec_path] || "spec"
-      @tmp_path = options[:tmp_path] || ENV['TMPDIR'] || "tmp"
+      @main_path = File.expand_path(options[:main_path] || "lib")
+      @spec_path = File.expand_path(options[:spec_path] || "spec")
+      @tmp_path = File.expand_path(options[:tmp_path] || ENV['TMPDIR'] || "tmp")
     end
 
     def spec_files
       FileList["#{spec_path}/**/*_spec.js"]
     end
+    
+    private 
   end
 end