+++ lib/shenandoah/tasks.rb
@@ -21,11 +21,15 @@ module Shenandoah
       create_run_task
     end
     
-    def run_specs
+    def run_specs(pattern=nil)
       files = @locator.spec_files
       if ENV['SHEN_SPEC']
         trace "limiting shenandoah specs based on #{ENV['SHEN_SPEC'].inspect}"
         files = files.select { |f| f =~ /#{ENV['SHEN_SPEC']}/ }
+      end
+      if pattern
+        trace "limiting shenandoah specs based on #{pattern.inspect}"
+        files = files.select { |f| f =~ /#{pattern}/ }
       end
       trace "running shenandoah specs\n - #{files.join("\n - ")}"
       successes = @runner.run_console(files)
@@ -64,8 +68,8 @@ module Shenandoah
     
     def create_run_task
       desc "Run the JavaScript specs"
-      task('shen:spec') do |t|
-        run_specs
+      task('shen:spec', :pattern) do |t, args|
+        run_specs args.pattern
       end
     end
   end