+++ ChangeLog.markdown
@@ -2,6 +2,7 @@
 =====
 
 * Correct bug in Shenandoah::Tasks which made plain-Rakefile use not work
+* Add task descriptions for all the tasks -- visible with rake -T
 
 0.1.0
 =====