+++ ChangeLog.markdown
@@ -4,6 +4,7 @@
 * Correct bug in Shenandoah::Tasks which made plain-Rakefile use not work
 * Add task descriptions for all the tasks -- visible with rake -T
 * Automatically absolutize locator paths for plain-Rakefile use
+* Allow specs to be constrained using rake task arguments instead of an env var
 
 0.1.0
 =====